+++ src/MarcRoute/MarcRoute.js
@@ -1,18 +1,19 @@
-import { useCallback } from 'react';
-import { Route } from 'react-router-dom';
+import {
+  Route,
+  useHistory,
+  useRouteMatch,
+} from 'react-router-dom';
 import PropTypes from 'prop-types';
 
 import { LoadingPane } from '@folio/stripes/components';
-import {
-  useStripes,
-  useUserTenantPermissions,
-} from '@folio/stripes/core';
+import { useStripes } from '@folio/stripes/core';
 
 import { QuickMarcEditorContainer } from '../QuickMarcEditor';
 import { applyCentralTenantInHeaders } from '../QuickMarcEditor/utils';
 import { QUICK_MARC_ACTIONS } from '../QuickMarcEditor/constants';
 import { QuickMarcProvider } from '../contexts';
-import { useIsShared } from '../hooks';
+import { getIsSharedFromUrl } from '../contexts/QuickMarcContext/utils';
+import { useCheckCentralTenantPermission } from '../hooks';
 
 const MarcRoute = ({
   externalRecordPath,
@@ -24,30 +25,33 @@ const MarcRoute = ({
   onSave,
 }) => {
   const stripes = useStripes();
+  const history = useHistory();
+  const match = useRouteMatch();
 
-  const { isShared } = useIsShared();
+  const {
+    instanceId,
+    externalId,
+  } = match;
+
+  const isShared = getIsSharedFromUrl(history.location.search);
 
   const {
     marcType,
     action,
   } = routeProps;
-  const centralTenantId = stripes.user.user?.consortium?.centralTenantId;
 
   const isRequestToCentralTenantFromMember = applyCentralTenantInHeaders(isShared, stripes, marcType)
     && action !== QUICK_MARC_ACTIONS.CREATE;
 
   const {
-    userPermissions: centralTenantPermissions,
-    isFetching: isCentralTenantPermissionsLoading,
-  } = useUserTenantPermissions({
-    tenantId: centralTenantId,
-  }, {
-    enabled: isRequestToCentralTenantFromMember,
+    isCentralTenantPermissionsLoading,
+    checkCentralTenantPerm,
+  } = useCheckCentralTenantPermission({
+    isShared,
+    marcType,
+    action,
+    enabled: true,
   });
-
-  const checkCentralTenantPerm = useCallback((perm) => {
-    return centralTenantPermissions.some(({ permissionName }) => permissionName === perm);
-  }, [centralTenantPermissions]);
 
   if (isCentralTenantPermissionsLoading) {
     return <LoadingPane />;
@@ -70,12 +74,15 @@ const MarcRoute = ({
           action={action}
           marcType={marcType}
           basePath={basePath}
+          isUsingRouter
         >
           <QuickMarcEditorContainer
             onClose={onClose}
             onSave={onSave}
             externalRecordPath={externalRecordPath}
             onCheckCentralTenantPerm={checkCentralTenantPerm}
+            instanceId={instanceId}
+            externalId={externalId}
           />
         </QuickMarcProvider>
       )}
@@ -87,7 +94,10 @@ MarcRoute.propTypes = {
   externalRecordPath: PropTypes.string.isRequired,
   path: PropTypes.string.isRequired,
   permission: PropTypes.string,
-  routeProps: PropTypes.object.isRequired,
+  routeProps: PropTypes.shape({
+    marcType: PropTypes.string.isRequired,
+    action: PropTypes.string.isRequired,
+  }).isRequired,
   basePath: PropTypes.string.isRequired,
   onClose: PropTypes.func.isRequired,
   onSave: PropTypes.func.isRequired,