+++ src/QuickMarcEditor/useSaveRecord/useSaveRecord.test.js
@@ -10,11 +10,13 @@ import {
 import { renderHook } from '@folio/jest-config-stripes/testing-library/react';
 
 import { QUICK_MARC_ACTIONS } from '../constants';
-import { ERROR_TYPES, MARC_TYPES } from '../../common';
+import {
+  ERROR_TYPES,
+  MARC_TYPES,
+} from '../../common';
 import {
   useAuthorityLinking,
   useValidation,
-  useIsShared,
 } from '../../hooks';
 import {
   useMarcRecordMutation,
@@ -53,11 +55,6 @@ jest.mock('../../hooks', () => ({
   ...jest.requireActual('../../hooks'),
   useAuthorityLinking: jest.fn(),
   useValidation: jest.fn((...params) => jest.requireActual('../../hooks').useValidation(...params)),
-  useIsShared: jest.fn().mockReturnValue({
-    isShared: false,
-    getIsShared: () => false,
-    setIsShared: jest.fn(),
-  }),
 }));
 
 jest.mock('../../queries', () => ({
@@ -1125,10 +1122,6 @@ describe('useSaveRecord', () => {
 
       beforeEach(() => {
         checkIfUserInCentralTenant.mockClear().mockReturnValue(true);
-        useIsShared.mockReturnValue({
-          isShared: false,
-          setIsShared: mockSetIsShared,
-        });
       });
 
       it('should mark record as shared', async () => {
@@ -1145,6 +1138,8 @@ describe('useSaveRecord', () => {
               action,
               marcType,
               continueAfterSave: { current: true },
+              isShared: false,
+              setIsShared: mockSetIsShared,
             },
             history,
           }),
@@ -1584,11 +1579,6 @@ describe('useSaveRecord', () => {
     describe('when a member tenant edits a shared record', () => {
       it('should apply the central tenant id for all authority linking ', async () => {
         checkIfUserInMemberTenant.mockClear().mockReturnValue(true);
-        useIsShared.mockReturnValue({
-          isShared: true,
-          getIsShared: () => true,
-          setIsShared: jest.fn(),
-        });
 
         const marcType = MARC_TYPES.BIB;
         const action = QUICK_MARC_ACTIONS.EDIT;
@@ -1604,6 +1594,8 @@ describe('useSaveRecord', () => {
               marcType,
               initialValues: getInitialValues(action, marcType),
               instance: getInstance(),
+              isShared: true,
+              setIsShared: jest.fn(),
             },
           }),
         });