+++ src/QuickMarcEditor/useSaveRecord/useSubmitRecord/useSumbitRecord.js
@@ -29,10 +29,7 @@ import {
 } from '../../utils';
 import getQuickMarcRecordStatus from '../../getQuickMarcRecordStatus';
 import { QuickMarcContext } from '../../../contexts';
-import {
-  useAuthorityLinking,
-  useIsShared,
-} from '../../../hooks';
+import { useAuthorityLinking } from '../../../hooks';
 import { useMarcRecordMutation } from '../../../queries';
 import { QUICK_MARC_ACTIONS } from '../../constants';
 
@@ -66,11 +63,11 @@ const useSubmitRecord = ({
     instance,
     continueAfterSave,
     relatedRecordVersion,
+    setIsShared,
   } = useContext(QuickMarcContext);
 
   const { actualizeLinks } = useAuthorityLinking({ marcType, action });
   const { updateMarcRecord } = useMarcRecordMutation({ tenantId });
-  const { setIsShared } = useIsShared();
 
   const redirectToRecord = useCallback(async (externalId, instanceId) => {
     if (marcType === MARC_TYPES.HOLDINGS) {