+++ src/contexts/QuickMarcContext/utils.js
@@ -1,0 +1,3 @@
+export const getIsSharedFromUrl = (searchQuery) => {
+  return new URLSearchParams(searchQuery).get('shared') === 'true';
+};