+++ src/hooks/index.js
@@ -1,4 +1,4 @@
 export * from './useAuthorityLinking';
 export * from './useSubfieldNavigation';
 export * from './useValidation';
-export * from './useIsShared';
+export * from './useCheckCentralTenantPermission';